Split game data between fusion/shared
