+++ src/mars_patcher/mf/starting.py
@@ -1,8 +1,9 @@
+from mars_patcher.constants.game_data import area_doors_ptrs, spriteset_ptrs
 from mars_patcher.mf.auto_generated_types import (
     MarsschemamfStartingitems,
     MarsschemamfStartinglocation,
 )
-from mars_patcher.mf.constants.game_data import area_doors_ptrs, spriteset_ptrs, starting_equipment
+from mars_patcher.mf.constants.game_data import starting_equipment
 from mars_patcher.mf.constants.items import BEAM_FLAGS, MISSILE_BOMB_FLAGS, SUIT_MISC_FLAGS
 from mars_patcher.mf.constants.reserved_space import ReservedConstants
 from mars_patcher.rom import Rom
