+++ src/mars_patcher/minimap.py
@@ -4,7 +4,7 @@ import json
 from typing import TYPE_CHECKING
 
 from mars_patcher.compress import comp_lz77, decomp_lz77
-from mars_patcher.mf.constants.game_data import minimap_ptrs
+from mars_patcher.constants.game_data import minimap_ptrs
 from mars_patcher.mf.data import get_data_path
 
 if TYPE_CHECKING:
