+++ src/mars_patcher/room_entry.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from mars_patcher.compress import comp_rle, decomp_rle
-from mars_patcher.mf.constants.game_data import area_room_entry_ptrs
+from mars_patcher.constants.game_data import area_room_entry_ptrs
 
 if TYPE_CHECKING:
     from types import TracebackType
